+++ test_stack_lts.sh
@@ -1,0 +1,4 @@
+set +e
+
+STACK_YAML=stack-7.10.yaml stack build  --no-terminal
+STACK_YAML=stack-8.0.yaml  stack build  --no-terminal